+++ contact/tests/test_views.py
@@ -13,7 +13,7 @@ from django.core.cache import cache
 from django.urls import reverse
 
 from contact import constants, forms, views
-from core.tests.helpers import create_response
+from core.tests.helpers import create_response, reload_urlconf
 
 
 def build_wizard_url(step):
@@ -591,7 +591,10 @@ def test_guidance_view_cms_retrieval(mock_lookup_by_slug, client):
 
 
 @mock.patch('directory_cms_client.client.cms_api_client.lookup_by_slug')
-def test_exporting_to_uk_cms_retrieval(mock_lookup_by_slug, client):
+def test_exporting_to_uk_cms_retrieval(mock_lookup_by_slug, client, settings):
+    settings.FEATURE_FLAGS['INTERNATIONAL_CONTACT_TRIAGE_ON'] = False
+    reload_urlconf()
+
     mock_lookup_by_slug.return_value = create_response()
 
     url = reverse(