+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre-release
 
 ### Implemented enhancements
+- XOT-1118 - use CMS heading for page title
 - XOT-1121 - add pagination to markets landing page, limit 12 per page
 - XOT-1042 - restyle markets grid and some elements of landing page
 - No ticket - Rename 'article' app folder to 'content'
@@ -36,6 +37,7 @@
 - XOT-1034 - Add new landing page, remove obsolete international page styling, update metadata for cms pages
 - TT-1808: Update directory components to add "no-validate" no cache middleware
 - No ticket - Add trade barriers to search
+- CI-501 - Added redirect from `/contact/triage/international/` to new great-international-ui url `/international/contact`
 - XOT-928 - Add user full name from new registration journey to SOO contact form
 
 ### Bug fixes